+++ src/content.js
@@ -12,7 +12,7 @@ const fileServer = require('./storage/file-server');
 const launchEnv = require('./launch-environment');
 const screenshot = require('./screenshot');
 const uptime = require('./uptime/uptime');
-const contentUptime = require('./content-uptime');
+const contentUptime = require('./uptime/content-uptime');
 const uptimeRendererHealth = require('./uptime/renderer-health');
 const scheduleParser = require('./scheduling/schedule-parser');
 const noViewerSchedulePlayer = require('./scheduling/schedule-player');
